cli, update readme
